requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.1
astroid==2.5.3
autopep8==1.5.6
backcall==0.2.0
blessings==1.7
boto3==1.17.28
botocore==1.20.33
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.6
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-allauth==0.44.0
django-configurations==2.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-model-utils==4.1.1
django-nose==1.4.7
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-storages==1.11.1
django-unique-upload==0.2.1
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==6.6.2
filelock==3.0.12
flake8==3.9.0
future==0.18.2
gunicorn==20.0.4
identify==2.2.3
idna==2.10
inflection==0.5.1
ipdb==0.13.6
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.8.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
lazy-object-proxy==1.6.0
livereload==2.6.3
lunr==0.5.8
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.1.2
mock==4.0.3
newrelic==6.2.0.156
nltk==3.6.1
nodeenv==1.6.0
nose==1.3.7
nose-progressive==1.5.2
oauthlib==3.1.0
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pre-commit==2.12.0
prompt-toolkit==3.0.17
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.0
Pygments==2.8.1
PyJWT==1.7.1
pylint==2.7.4
pylint-django==2.0.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.3
wcwidth==0.2.5
wrapt==1.12.1
